refactor: separate orm tables
